+++ tests/smoke/test_import.py
@@ -1,4 +1,3 @@
 def test_import_pycrmkit() -> None:
     import pycrmkit
-
-    assert pycrmkit.__version__ == "0.4.0b2"
+    assert pycrmkit.__version__ == "0.4.0rc1"